fix: Query문의 SUM() type에 맞게 dto type 수정

+++ src/main/java/Hampouch/server/domain/battle/dto/BattleDetailResponse.java
@@ -36,8 +36,8 @@ public record BattleDetailResponse(
             String nickname,
             String avatarUrl,
             Integer rank,
-            int todayAmount,
-            int totalAmount,
+            long todayAmount,
+            long totalAmount,
             boolean isValid
     ) {
     }
